Moved common functions to AESmath
Moved common math functions (galois field mathematics) and key expansion functions to AESmath
Cleaned up the loops in keyExpansion
Reformatted curly brackets for consistency
